requirements.txt
fastapi
uvicorn[standard]
tensorflow-macos
tensorflow-metal
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
ImageIO==2.37.2
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2026.1.14
tzdata==2025.3
xgboost==3.1.3
